Added name and group attribute to Dining class. Modified code in the add button event to reflect these attributes being added. Added a void method and HasMap attribute to main GUI class to remove food groups, not yet utilized.
