+++ src/team4/personaldietary/Dining.java
@@ -3,9 +3,29 @@ package team4.personaldietary;
 public class Dining {
 
     private String time;
+    private String group;
+    private String name;
 
-    public Dining(String time) {
+    public Dining(String time, String group, String name) {
         this.time = time;
+        this.group = group;
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getGroup() {
+        return group;
+    }
+
+    public void setGroup(String group) {
+        this.group = group;
     }
 
     public String getTime() {
@@ -25,7 +45,7 @@ public class Dining {
     }
 
     public String toString() {
-        return "" +time;
+        return "Name: "+name+" Group: "+group+" Time: " +time;
     }
 
 }
